[SkG] Renames and better binding
